Refactor create, update, delete methods from IntacctClient into new IntacctObjectTrait, update RequestHandler accordingly, and added new IaClass under Dimension to test IntacctObjectTrait methods.
