+++ test/Intacct/Dimension/IaClassTest.php
@@ -1,0 +1,539 @@
+<?php
+
+/*
+ * Copyright 2016 Intacct Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"). You may not
+ * use this file except in compliance with the License. You may obtain a copy
+ * of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * or in the "LICENSE" file accompanying this file. This file is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+namespace Intacct\Dimension;
+
+use Intacct\IntacctClient;
+use GuzzleHttp\Psr7\Response;
+use GuzzleHttp\Handler\MockHandler;
+use Intacct\Xml\Request\Operation\Content\Record;
+
+class IaClassTest extends \PHPUnit_Framework_TestCase
+{
+
+    /**
+     *
+     * @var IntacctClient
+     */
+    private $client;
+
+    /**
+     * Sets up the fixture, for example, opens a network connection.
+     * This method is called before a test is executed.
+     */
+    protected function setUp()
+    {
+        //the IntacctClient constructor will always get a session id, so mock it
+        $xml = <<<EOF
+<?xml version="1.0" encoding="UTF-8"?>
+<response>
+      <control>
+            <status>success</status>
+            <senderid>testsenderid</senderid>
+            <controlid>sessionProvider</controlid>
+            <uniqueid>false</uniqueid>
+            <dtdversion>3.0</dtdversion>
+      </control>
+      <operation>
+            <authentication>
+                  <status>success</status>
+                  <userid>testuser</userid>
+                  <companyid>testcompany</companyid>
+                  <sessiontimestamp>2015-12-06T15:57:08-08:00</sessiontimestamp>
+            </authentication>
+            <result>
+                  <status>success</status>
+                  <function>getAPISession</function>
+                  <controlid>getSession</controlid>
+                  <data>
+                        <api>
+                              <sessionid>testSeSsionID..</sessionid>
+                              <endpoint>https://p1.intacct.com/ia/xml/xmlgw.phtml</endpoint>
+                        </api>
+                  </data>
+            </result>
+      </operation>
+</response>
+EOF;
+        $headers = [
+            'Content-Type' => 'text/xml; encoding="UTF-8"',
+        ];
+        $mockResponse = new Response(200, $headers, $xml);
+        $mock = new MockHandler([
+            $mockResponse,
+        ]);
+
+        $this->client = new IntacctClient([
+            'sender_id' => 'testsenderid',
+            'sender_password' => 'pass123!',
+            'company_id' => 'testcompany',
+            'user_id' => 'testuser',
+            'user_password' => 'testpass',
+            'mock_handler' => $mock,
+        ]);
+    }
+
+    /**
+     * Tears down the fixture, for example, closes a network connection.
+     * This method is called after a test is executed.
+     */
+    protected function tearDown()
+    {
+
+    }
+
+    /**
+     * @covers Intacct\Dimension\IaClass::create
+     * @covers Intacct\Xml\RequestHandler::executeContent
+     * @covers Intacct\IntacctClient::getSessionConfig
+     */
+    public function testCreateSuccess()
+    {
+        $xml = <<<EOF
+<?xml version="1.0" encoding="UTF-8"?>
+<response>
+      <control>
+            <status>success</status>
+            <senderid>testsenderid</senderid>
+            <controlid>requestControlId</controlid>
+            <uniqueid>false</uniqueid>
+            <dtdversion>3.0</dtdversion>
+      </control>
+      <operation>
+            <authentication>
+                  <status>success</status>
+                  <userid>testuser</userid>
+                  <companyid>testcompany</companyid>
+                  <sessiontimestamp>2016-01-24T14:26:56-08:00</sessiontimestamp>
+            </authentication>
+            <result>
+                  <status>success</status>
+                  <function>create</function>
+                  <controlid>create</controlid>
+                  <data listtype="objects" count="2">
+                        <class>
+                              <RECORDNO>5</RECORDNO>
+                              <CLASSID>UT01</CLASSID>
+                        </class>
+                        <class>
+                              <RECORDNO>6</RECORDNO>
+                              <CLASSID>UT02</CLASSID>
+                        </class>
+                  </data>
+            </result>
+      </operation>
+</response>
+EOF;
+        $headers = [
+            'Content-Type' => 'text/xml; encoding="UTF-8"',
+        ];
+        $mockResponse = new Response(200, $headers, $xml);
+        $mock = new MockHandler([
+            $mockResponse,
+        ]);
+
+        $create = [
+            'records' => [
+                new Record([
+                    'object' => 'CLASS',
+                    'fields' => [
+                        'CLASSID' => 'UT01',
+                        'NAME' => 'Unit Test 01',
+                    ],
+                ]),
+                new Record([
+                    'object' => 'CLASS',
+                    'fields' => [
+                        'CLASSID' => 'UT02',
+                        'NAME' => 'Unit Test 02',
+                    ],
+                ]),
+            ],
+            'mock_handler' => $mock,
+        ];
+        $data = $this->client->dimensions->iaClass->create($create);
+        $this->assertEquals($data->getStatus(), 'success');
+        $this->assertEquals($data->getFunction(), 'create');
+        $this->assertEquals($data->getControlId(), 'create');
+    }
+
+    /**
+     * @covers Intacct\Dimension\IaClass::create
+     * @covers Intacct\Xml\RequestHandler::executeContent
+     * @covers Intacct\IntacctClient::getSessionConfig
+     * @expectedException \Intacct\Xml\Response\Operation\ResultException
+     * @expectedExceptionMessage An error occurred trying to create records
+     */
+    public function testCreateFailure()
+    {
+        $xml = <<<EOF
+<?xml version="1.0" encoding="UTF-8"?>
+<response>
+      <control>
+            <status>success</status>
+            <senderid>testsenderid</senderid>
+            <controlid>requestControlId</controlid>
+            <uniqueid>false</uniqueid>
+            <dtdversion>3.0</dtdversion>
+      </control>
+      <operation>
+            <authentication>
+                  <status>success</status>
+                  <userid>testuser</userid>
+                  <companyid>testcompany</companyid>
+                  <sessiontimestamp>2016-01-24T14:26:56-08:00</sessiontimestamp>
+            </authentication>
+            <result>
+                  <status>failure</status>
+                  <function>create</function>
+                  <controlid>create</controlid>
+                  <data listtype="objects" count="0"/>
+                  <errormessage>
+                        <error>
+                              <errorno>BL34000061</errorno>
+                              <description></description>
+                              <description2>Another Class with the given value(s) UT01  already exists</description2>
+                              <correction>Use a unique value instead.</correction>
+                        </error>
+                        <error>
+                              <errorno>BL01001973</errorno>
+                              <description></description>
+                              <description2>Could not create class record!</description2>
+                              <correction></correction>
+                        </error>
+                        <error>
+                              <errorno>BL01001973</errorno>
+                              <description></description>
+                              <description2>Could not create Class record!</description2>
+                              <correction></correction>
+                        </error>
+                  </errormessage>
+            </result>
+      </operation>
+</response>
+EOF;
+        $headers = [
+            'Content-Type' => 'text/xml; encoding="UTF-8"',
+        ];
+        $mockResponse = new Response(200, $headers, $xml);
+        $mock = new MockHandler([
+            $mockResponse,
+        ]);
+
+        $create = [
+            'records' => [
+                new Record([
+                    'object' => 'CLASS',
+                    'fields' => [
+                        'CLASSID' => 'UT01',
+                        'NAME' => 'Unit Test 01',
+                    ],
+                ]),
+                new Record([
+                    'object' => 'CLASS',
+                    'fields' => [
+                        'CLASSID' => 'UT02',
+                        'NAME' => 'Unit Test 02',
+                    ],
+                ]),
+            ],
+            'mock_handler' => $mock,
+        ];
+        $this->client->dimensions->iaClass->create($create);
+    }
+
+    /**
+     * @covers Intacct\Dimension\IaClass::update
+     * @covers Intacct\Xml\RequestHandler::executeContent
+     * @covers Intacct\IntacctClient::getSessionConfig
+     */
+    public function testUpdateSuccess()
+    {
+        $xml = <<<EOF
+<?xml version="1.0" encoding="UTF-8"?>
+<response>
+      <control>
+            <status>success</status>
+            <senderid>testsenderid</senderid>
+            <controlid>requestControlId</controlid>
+            <uniqueid>false</uniqueid>
+            <dtdversion>3.0</dtdversion>
+      </control>
+      <operation>
+            <authentication>
+                  <status>success</status>
+                  <userid>testuser</userid>
+                  <companyid>testcompany</companyid>
+                  <sessiontimestamp>2016-01-24T14:26:56-08:00</sessiontimestamp>
+            </authentication>
+            <result>
+                  <status>success</status>
+                  <function>update</function>
+                  <controlid>update</controlid>
+                  <data listtype="objects" count="2">
+                        <class>
+                              <RECORDNO>5</RECORDNO>
+                              <CLASSID>UT01</CLASSID>
+                        </class>
+                        <class>
+                              <RECORDNO>6</RECORDNO>
+                              <CLASSID>UT02</CLASSID>
+                        </class>
+                  </data>
+            </result>
+      </operation>
+</response>
+EOF;
+        $headers = [
+            'Content-Type' => 'text/xml; encoding="UTF-8"',
+        ];
+        $mockResponse = new Response(200, $headers, $xml);
+        $mock = new MockHandler([
+            $mockResponse,
+        ]);
+
+        $update = [
+            'records' => [
+                new Record([
+                    'object' => 'CLASS',
+                    'fields' => [
+                        'CLASSID' => 'UT01',
+                        'NAME' => 'Unit Test 01',
+                    ],
+                ]),
+                new Record([
+                    'object' => 'CLASS',
+                    'fields' => [
+                        'CLASSID' => 'UT02',
+                        'NAME' => 'Unit Test 02',
+                    ],
+                ]),
+            ],
+            'mock_handler' => $mock,
+        ];
+        $data = $this->client->dimensions->iaClass->update($update);
+        $this->assertEquals($data->getStatus(), 'success');
+        $this->assertEquals($data->getFunction(), 'update');
+        $this->assertEquals($data->getControlId(), 'update');
+    }
+
+    /**
+     * @covers Intacct\Dimension\IaClass::update
+     * @covers Intacct\Xml\RequestHandler::executeContent
+     * @covers Intacct\IntacctClient::getSessionConfig
+     * @expectedException \Intacct\Xml\Response\Operation\ResultException
+     * @expectedExceptionMessage An error occurred trying to update records
+     */
+    public function testUpdateFailure()
+    {
+        $xml = <<<EOF
+<?xml version="1.0" encoding="UTF-8"?>
+<response>
+      <control>
+            <status>success</status>
+            <senderid>testsenderid</senderid>
+            <controlid>requestControlId</controlid>
+            <uniqueid>false</uniqueid>
+            <dtdversion>3.0</dtdversion>
+      </control>
+      <operation>
+            <authentication>
+                  <status>success</status>
+                  <userid>testuser</userid>
+                  <companyid>testcompany</companyid>
+                  <sessiontimestamp>2016-01-24T14:26:56-08:00</sessiontimestamp>
+            </authentication>
+            <result>
+                  <status>failure</status>
+                  <function>update</function>
+                  <controlid>create</controlid>
+                  <data listtype="objects" count="1">
+                        <class>
+                              <RECORDNO>5</RECORDNO>
+                              <CLASSID>UT01</CLASSID>
+                        </class>
+                  </data>
+                  <errormessage>
+                        <error>
+                              <errorno>Cannot update non-existing CLASS with CLASSID=UT99.</errorno>
+                              <description></description>
+                              <description2></description2>
+                              <correction></correction>
+                        </error>
+                  </errormessage>
+            </result>
+      </operation>
+</response>
+EOF;
+        $headers = [
+            'Content-Type' => 'text/xml; encoding="UTF-8"',
+        ];
+        $mockResponse = new Response(200, $headers, $xml);
+        $mock = new MockHandler([
+            $mockResponse,
+        ]);
+
+        $update = [
+            'records' => [
+                new Record([
+                    'object' => 'CLASS',
+                    'fields' => [
+                        'CLASSID' => 'UT01',
+                        'NAME' => 'Unit Test 01',
+                    ],
+                ]),
+                new Record([
+                    'object' => 'CLASS',
+                    'fields' => [
+                        'CLASSID' => 'UT99',
+                        'NAME' => 'Unit Test 99',
+                    ],
+                ]),
+            ],
+            'mock_handler' => $mock,
+        ];
+        $this->client->dimensions->iaClass->update($update);
+    }
+
+
+    /**
+     * @covers Intacct\Dimension\IaClass::delete
+     * @covers Intacct\Xml\RequestHandler::executeContent
+     * @covers Intacct\IntacctClient::getSessionConfig
+     */
+    public function testDeleteSuccess()
+    {
+        $xml = <<<EOF
+<?xml version="1.0" encoding="UTF-8"?>
+<response>
+      <control>
+            <status>success</status>
+            <senderid>testsenderid</senderid>
+            <controlid>requestControlId</controlid>
+            <uniqueid>false</uniqueid>
+            <dtdversion>3.0</dtdversion>
+      </control>
+      <operation>
+            <authentication>
+                  <status>success</status>
+                  <userid>testuser</userid>
+                  <companyid>testcompany</companyid>
+                  <sessiontimestamp>2016-01-24T14:26:56-08:00</sessiontimestamp>
+            </authentication>
+            <result>
+                  <status>success</status>
+                  <function>delete</function>
+                  <controlid>delete</controlid>
+            </result>
+      </operation>
+</response>
+EOF;
+        $headers = [
+            'Content-Type' => 'text/xml; encoding="UTF-8"',
+        ];
+        $mockResponse = new Response(200, $headers, $xml);
+        $mock = new MockHandler([
+            $mockResponse,
+        ]);
+
+        $delete = [
+            'object' => 'CLASS',
+            'keys' => [
+                '5',
+                '6',
+            ],
+            'mock_handler' => $mock,
+        ];
+        $data = $this->client->dimensions->iaClass->delete($delete);
+        $this->assertEquals($data->getStatus(), 'success');
+        $this->assertEquals($data->getFunction(), 'delete');
+        $this->assertEquals($data->getControlId(), 'delete');
+    }
+
+    /**
+     * @covers Intacct\Dimension\IaClass::delete
+     * @covers Intacct\Xml\RequestHandler::executeContent
+     * @covers Intacct\IntacctClient::getSessionConfig
+     * @expectedException \Intacct\Xml\Response\Operation\ResultException
+     * @expectedExceptionMessage An error occurred trying to delete records
+     */
+    public function testDeleteFailure()
+    {
+        $xml = <<<EOF
+<?xml version="1.0" encoding="UTF-8"?>
+<response>
+      <control>
+            <status>success</status>
+            <senderid>testsenderid</senderid>
+            <controlid>requestControlId</controlid>
+            <uniqueid>false</uniqueid>
+            <dtdversion>3.0</dtdversion>
+      </control>
+      <operation>
+            <authentication>
+                  <status>success</status>
+                  <userid>testuser</userid>
+                  <companyid>testcompany</companyid>
+                  <sessiontimestamp>2016-01-24T14:26:56-08:00</sessiontimestamp>
+            </authentication>
+            <result>
+                  <status>failure</status>
+                  <function>delete</function>
+                  <controlid>delete</controlid>
+                  <errormessage>
+                        <error>
+                              <errorno>BL01001973</errorno>
+                              <description></description>
+                              <description2>Cannot find class with key &#039;UT01&#039; to delete.</description2>
+                              <correction></correction>
+                        </error>
+                  </errormessage>
+            </result>
+      </operation>
+</response>
+EOF;
+        $headers = [
+            'Content-Type' => 'text/xml; encoding="UTF-8"',
+        ];
+        $mockResponse = new Response(200, $headers, $xml);
+        $mock = new MockHandler([
+            $mockResponse,
+        ]);
+
+        $delete = [
+            'object' => 'CLASS',
+            'keys' => [
+                'UT01',
+                'UT02',
+            ],
+            'mock_handler' => $mock,
+        ];
+        $this->client->dimensions->iaClass->delete($delete);
+    }
+
+    /**
+     * @covers Intacct\Dimension\IaClass::inspect
+     * @todo   Implement testInspect().
+     */
+    public function testInspect()
+    {
+        // Remove the following lines when you implement this test.
+        $this->markTestIncomplete(
+            'This test has not been implemented yet.'
+        );
+    }
+}
